refactor: move parse tree printing into a new file
